Added in support for relative fields - if a form has two fields where one value depends on another you can now leave one of those two fields empty and it will be filled on submit.

+++ backend/db/dbutils.js
@@ -182,7 +182,7 @@ function applyFieldsFilter(doc, fields) {
                 break;
 
             case 'number':
-                doc[field.key] = parseInt(value);
+                doc[field.key] = parseFloat(value);
                 break;
 
             case 'date':
